% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{make_project}
\alias{make_project}
\title{Create a now project directory}
\usage{
make_project(path, ref = TRUE)
}
\arguments{
\item{path}{Absolute or relative project path. If the directory does not exist, it will be created.
Spaces will be replaced by a graceful underscore \verb{_} in the files names.
Note that it is a (very) good practice to have no white space or weird characters in your project name}

\item{ref}{A boolean. if \code{TRUE}, a References folder will be created to hold the literature related to your project}
}
\value{
A beautifully organised project folder
}
\description{
Create a now project directory
}
\examples{
\dontrun{
make_project("kickassAnalysis")
}
}
