% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeproject-package.R
\docType{package}
\name{makeproject-package}
\alias{makeproject}
\alias{makeproject-package}
\title{makeproject: Creates A Tidy Working Space for a New Project}
\description{
This (very small) package creates a nicely organized projetc folder.
  Probably more convenient functions will be added in the future to manage the projects.
}
\author{
\strong{Maintainer}: Matthieu Mulot \email{matthieu@lebiostatisticien.fr}

}
\keyword{internal}
