% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageproject-package.R
\docType{package}
\name{manageproject-package}
\alias{manageproject}
\alias{manageproject-package}
\title{manageproject: Creates A Tidy Working Space for a New Project}
\description{
This (very small) package creates a nicely organized project folder.
  Probably more convenient functions will be added in the future to manage the projects.
}
\author{
\strong{Maintainer}: Matthieu Mulot \email{matthieu@lebiostatisticien.fr}

}
\keyword{internal}
